work

+++ examples/batch_nvprof/main.go
@@ -126,7 +126,7 @@ func main() {
 		options.Device(device, 0),
 		options.Graph([]byte(graph)),
 		options.Weights([]byte(weights)),
-		options.BatchSize(uint32(batchSize)))
+		options.BatchSize(batchSize))
 	if err != nil {
 		panic(err)
 	}
